add anteHandler for verfying report submiter and feedId (#62)
* add anteHandler for verfying report submiter and feedId

* minor fix
