+++ x/chainlink/ante/ante.go
@@ -38,6 +38,7 @@ func NewAnteHandler(
 			// all customized anteHandler below
 			NewModuleOwnerDecorator(chainLinkKeeper),
 			NewFeedDecorator(chainLinkKeeper),
+			NewFeedDataDecorator(chainLinkKeeper),
 		)
 
 		return anteHandler(ctx, tx, sim)
@@ -124,6 +125,39 @@ func (fd FeedDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, ne
 			feed := fd.chainLinkKeeper.GetFeed(ctx, t.GetFeedId())
 			if !feed.Feed.Empty() {
 				return ctx, sdkerrors.Wrap(sdkerrors.ErrUnknownRequest, "feed already exists")
+			}
+		default:
+			continue
+		}
+	}
+
+	return next(ctx, tx, simulate)
+}
+
+type FeedDataDecorator struct {
+	chainLinkKeeper chainlinkkeeper.Keeper
+}
+
+func NewFeedDataDecorator(chainLinkKeeper chainlinkkeeper.Keeper) FeedDataDecorator {
+	return FeedDataDecorator{
+		chainLinkKeeper: chainLinkKeeper,
+	}
+}
+
+func (fd FeedDataDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, next sdk.AnteHandler) (newCtx sdk.Context, err error) {
+	if len(tx.GetMsgs()) == 0 {
+		return ctx, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "invalid Msg: empty Msg: %T", tx)
+	}
+
+	for _, msg := range tx.GetMsgs() {
+		switch t := msg.(type) {
+		case *types.MsgFeedData:
+			feed := fd.chainLinkKeeper.GetFeed(ctx, t.GetFeedId())
+			if feed.Feed.Empty() {
+				return ctx, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, "feed not exist")
+			}
+			if !(types.DataProviders)(feed.GetFeed().GetDataProviders()).Contains(t.GetSubmitter()) {
+				return ctx, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "invalid data provider")
 			}
 		default:
 			continue
