+++ x/chainlink/client/cli/txModuleInfo.go
@@ -1,11 +1,6 @@
 package cli
 
 import (
-	"errors"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"strconv"
-	"strings"
-
 	"github.com/ChainSafe/chainlink-cosmos/x/chainlink/types"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
@@ -78,71 +73,3 @@ func CmdTransferModuleOwnership() *cobra.Command {
 	return cmd
 }
 
-func CmdAddFeed() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "addFeed [feedId] [feedOwnerAddress] [submissionCount] [heartbeatTrigger] [deviationThresholdTrigger] [initDataProviderList]",
-		Short: "Add new feed. Signer must be the existing module owner.",
-		Args:  cobra.MinimumNArgs(6),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			argsFeedId := strings.TrimSpace(args[0])
-			argsFeedOwnerAddr := strings.TrimSpace(args[1])
-			argsSubmissionCount := strings.TrimSpace(args[2])
-			argsHeartbeatTrigger := strings.TrimSpace(args[3])
-			argsDeviationThresholdTrigger := strings.TrimSpace(args[4])
-			argsInitDataProviderListStr := strings.TrimSpace(args[5])
-
-			submissionCount, err := strconv.Atoi(argsSubmissionCount)
-			if err != nil {
-				return err
-			}
-			heartbeatTrigger, err := strconv.Atoi(argsHeartbeatTrigger)
-			if err != nil {
-				return err
-			}
-			deviationThresholdTrigger, err := strconv.Atoi(argsDeviationThresholdTrigger)
-			if err != nil {
-				return err
-			}
-
-			argsInitDataProviderList := strings.Split(strings.TrimSpace(argsInitDataProviderListStr), " ")
-			if len(argsInitDataProviderList)%2 != 0 {
-				return errors.New("invalid init data provider pairs")
-			}
-
-			initDataProviderList := make([]*types.DataProvider, 0, len(argsInitDataProviderList)/2)
-			i := 0
-			for i < len(argsInitDataProviderList) {
-				addr, err := sdk.AccAddressFromBech32(strings.TrimSpace(argsInitDataProviderList[i]))
-				if err != nil {
-					return sdkerrors.Wrapf(err, "invalid init data provider address: %s", argsInitDataProviderList[i])
-				}
-
-				initDataProviderList = append(initDataProviderList, &types.DataProvider{
-					Address: addr,
-					PubKey:  []byte(strings.TrimSpace(argsInitDataProviderList[i+1])),
-				})
-				i = i + 2
-			}
-
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			feedOwnerAddr, err := sdk.AccAddressFromBech32(argsFeedOwnerAddr)
-			if err != nil {
-				return err
-			}
-
-			msg := types.NewMsgFeed(argsFeedId, feedOwnerAddr, clientCtx.GetFromAddress(), initDataProviderList, uint32(submissionCount), uint32(heartbeatTrigger), uint32(deviationThresholdTrigger))
-			if err := msg.ValidateBasic(); err != nil {
-				return err
-			}
-			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
-		},
-	}
-
-	flags.AddTxFlagsToCmd(cmd)
-
-	return cmd
-}
